Bug Hunting

+++ facedetector/HistogramExtractor.cpp
@@ -10,7 +10,7 @@
 std::shared_ptr<ProcessedImage> HistogramExtractor::ProcessImage(cv::Mat image) {
     ProcessedImage* result = new ProcessedImage();
     result->histogramCount = (image.cols - (AREA_SIZE - 1)) * (image.rows - (AREA_SIZE - 1));  //POCET HISTOGRAMU V OBRAZKU
-    result->histograms = new int[result->histogramCount * HIST_SIZE];
+    result->histograms = new short[result->histogramCount * HIST_SIZE];
     this->extractHistograms(result->histograms, result->histogramCount, image.data, image.cols, image.rows);
 
     return std::shared_ptr<ProcessedImage>(result);
