+++ facedetector/ProcessedImage.h
@@ -9,7 +9,7 @@
 class ProcessedImage {
 public:
     int histogramCount;
-    int* histograms;
+    short* histograms;
 };
